Move quoting the label values out of LabeledMetric construction
* Move quoting values out of LabeledMetric construction

Quoting the value of the labels is computationally heavy compared to constructing a LabeledMetric. It can be deferred until after the construction of an ObjectName out of the labeled metric name which its result is cached.
